chore: rename old API
